refactor: view 패키지 생성
