+++ app/src/main/java/com/sdhong/pokemonapp/view/HistoryFragment.kt
@@ -1,4 +1,4 @@
-package com.sdhong.pokemonapp
+package com.sdhong.pokemonapp.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,15 +6,14 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
+import com.sdhong.pokemonapp.GridSpacingItemDecoration
 import com.sdhong.pokemonapp.common.Grid
-import com.sdhong.pokemonapp.databinding.FragmentAllPokemonBinding
-import com.sdhong.pokemonapp.model.Pokemon
-import com.sdhong.pokemonapp.model.allPokemons
+import com.sdhong.pokemonapp.databinding.FragmentHistoryBinding
 import com.sdhong.pokemonapp.model.seenPokemons
 
-class AllPokemonFragment : Fragment() {
+class HistoryFragment : Fragment() {
 
-    private lateinit var binding: FragmentAllPokemonBinding
+    private lateinit var binding: FragmentHistoryBinding
     private val mainAdapter = MainAdapter()
 
     override fun onCreateView(
@@ -22,14 +21,14 @@ class AllPokemonFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentAllPokemonBinding.inflate(inflater, container, false)
+        binding = FragmentHistoryBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.recyclerViewAllPokemon.run {
+        binding.recyclerViewHistory.run {
             adapter = mainAdapter
             layoutManager = GridLayoutManager(context, Grid.DEFAULT_SPAN_COUNT)
             addItemDecoration(
@@ -41,21 +40,17 @@ class AllPokemonFragment : Fragment() {
             )
         }
 
-        mainAdapter.submitList(allPokemons)
         mainAdapter.setOnClick(::onPokemonClick)
     }
 
-    fun onPokemonClick(position: Int) {
-        val pokemon = allPokemons[position]
-        seenPokemons.add(
-            0,
-            Pokemon.Seen(
-                name = pokemon.name,
-                imgUrl = pokemon.imgUrl,
-                lastViewed = System.currentTimeMillis().toString()
-            )
-        )
+    override fun onResume() {
+        super.onResume()
 
+        mainAdapter.submitList(seenPokemons.toList())
+    }
+
+    fun onPokemonClick(position: Int) {
+        val pokemon = seenPokemons[position]
         // TODO: 상세 페이지 이동
     }
 }
