+++ app/src/main/java/com/sdhong/pokemonapp/view/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.sdhong.pokemonapp
+package com.sdhong.pokemonapp.view
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -7,6 +7,7 @@ import androidx.appcompat.content.res.AppCompatResources
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.google.android.material.tabs.TabLayoutMediator
+import com.sdhong.pokemonapp.ViewPagerAdapter
 import com.sdhong.pokemonapp.databinding.ActivityMainBinding
 import com.sdhong.pokemonapp.model.MainTab
