+++ app/src/main/res/layout/fragment_all_pokemon.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AllPokemonFragment">
+    tools:context=".view.AllPokemonFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewAllPokemon"
